requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cached-property==1.5.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==5.5
cryptography==44.0.2
decorator==5.2.1
dill==0.3.9
dotify==2.0.6
exejs==0.0.4
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
greenlet==3.1.1
h11==0.14.0
html5==0.0.9
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
itsdangerous==2.2.0
jh2==5.0.8
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.3.2
magic-filter==1.0.12
Markdown==3.8
MarkupSafe==3.0.2
moviepy==2.1.2
multidict==6.3.2
multiprocess==0.70.17
mutagen==1.47.0
niquests==3.14.0
numpy==2.2.5
packaging==24.2
pathos==0.3.3
pefile==2023.2.7
pillow==10.4.0
pox==0.3.5
ppft==1.7.6.9
proglog==0.1.11
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dotenv==1.1.0
python-jsonschema-objects==0.5.7
python-telegram-bot==22.0
pytube==15.0.0
pywin32-ctypes==0.2.3
qh3==1.4.4
ratelim==0.1.6
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
spotipy==2.25.1
SQLAlchemy==2.0.39
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
translator==0.0.9
translators==5.9.9
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
urllib3-future==2.12.917
wassima==1.2.2
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
youtube-search-python==1.6.6
